# doc/api.yaml
openapi: 3.0.0
paths:
  /:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
  /user:
    get:
      operationId: UsersController_getAllUsers
      summary: Get all users
      description: Gets all users
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseUserDto'
      tags: &ref_0
        - Users
    post:
      operationId: UsersController_createUser
      summary: Create user
      description: Creates a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The user has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUserDto'
        '400':
          description: Bad request. body does not contain required fields
      tags: *ref_0
  /user/{id}:
    get:
      operationId: UsersController_getUserById
      summary: Get single user by id
      description: Get single user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: User ID (UUID v4)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUserDto'
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User not found
      tags: *ref_0
    put:
      operationId: UsersController_updatePassword
      summary: Update a user's password
      description: Updates a user's password by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: User ID (UUID v4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: The user has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUserDto'
        '400':
          description: Bad request. userId is invalid (not uuid)
        '403':
          description: oldPassword is wrong
        '404':
          description: User not found
      tags: *ref_0
    delete:
      operationId: UsersController_deleteUser
      summary: Delete user
      description: Deletes user by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: User ID (UUID v4)
      responses:
        '204':
          description: The user has been deleted
        '400':
          description: Bad request. userId is invalid (not uuid)
        '404':
          description: User not found
      tags: *ref_0
  /artist:
    get:
      operationId: ArtistsController_getAllArtists
      summary: Get all artists
      description: Gets all artists
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
      tags: &ref_1
        - Artists
    post:
      operationId: ArtistsController_createArtist
      summary: Create artist
      description: Creates a new artist
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistDto'
      responses:
        '201':
          description: The artist has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. body does not contain required fields
      tags: *ref_1
  /artist/{id}:
    get:
      operationId: ArtistsController_getArtistById
      summary: Get single artist by id
      description: Get single artist by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Artist ID (UUID v4)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags: *ref_1
    put:
      operationId: ArtistsController_updateArtist
      summary: Update artist information
      description: Update library artist information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Artist ID (UUID v4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistDto'
      responses:
        '200':
          description: The artist has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags: *ref_1
    delete:
      operationId: ArtistsController_deleteArtist
      summary: Delete artist
      description: Deletes artist by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Artist ID (UUID v4)
      responses:
        '204':
          description: The artist has been deleted
        '400':
          description: Bad request. artistId is invalid (not uuid)
        '404':
          description: Artist not found
      tags: *ref_1
  /album:
    get:
      operationId: AlbumsController_getAllArtists
      summary: Get all albums
      description: Gets all albums
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Album'
      tags: &ref_2
        - Albums
    post:
      operationId: AlbumsController_createArtist
      summary: Create album
      description: Creates a new album
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumDto'
      responses:
        '201':
          description: The album has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. body does not contain required fields
      tags: *ref_2
  /album/{id}:
    get:
      operationId: AlbumsController_getArtistById
      summary: Get single album by id
      description: Get single user by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Album ID (UUID v4)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      tags: *ref_2
    put:
      operationId: AlbumsController_updateArtist
      summary: Update album information
      description: Update library album information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Album ID (UUID v4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlbumDto'
      responses:
        '200':
          description: The album has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      tags: *ref_2
    delete:
      operationId: AlbumsController_deleteArtist
      summary: Delete album
      description: Deletes album by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Album ID (UUID v4)
      responses:
        '204':
          description: The album has been deleted
        '400':
          description: Bad request. albumId is invalid (not uuid)
        '404':
          description: Album not found
      tags: *ref_2
  /track:
    get:
      operationId: TracksController_getAllTracks
      summary: Get all tracks
      description: Gets all tracks
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
      tags: &ref_3
        - Tracks
    post:
      operationId: TracksController_createArtist
      summary: Create track
      description: Creates a new track
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackDto'
      responses:
        '201':
          description: The track has been created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. body does not contain required fields
      tags: *ref_3
  /track/{id}:
    get:
      operationId: TracksController_getTrackById
      summary: Get single track by id
      description: Get single track by id
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Track ID (UUID v4)
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags: *ref_3
    put:
      operationId: TracksController_updateTrack
      summary: Update track information
      description: Update library track information by UUID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Track ID (UUID v4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackDto'
      responses:
        '200':
          description: The track has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags: *ref_3
    delete:
      operationId: TracksController_deleteTrack
      summary: Delete track
      description: Deletes track by ID
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
          description: Track ID (UUID v4)
      responses:
        '204':
          description: The track has been deleted
        '400':
          description: Bad request. trackId is invalid (not uuid)
        '404':
          description: Track not found
      tags: *ref_3
info:
  title: Home Library Service
  description: Home music library service
  version: 1.0.0
  contact: {}
tags: []
servers: []
components:
  schemas:
    ResponseUserDto:
      type: object
      properties:
        login:
          type: string
          example: TestUser
        version:
          type: number
          example: 1
        createdAt:
          type: number
          example: 1655000000
        updatedAt:
          type: number
          example: 1655000000
      required:
        - login
        - version
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          example: TestUser
        password:
          type: string
          example: secret_password
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          example: secret_password
          description: The user's old password
        newPassword:
          type: string
          example: secret_password2
          description: The user's new password
      required:
        - oldPassword
        - newPassword
    Artist:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        name:
          type: string
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
      required:
        - id
        - name
        - grammy
    ArtistDto:
      type: object
      properties:
        name:
          type: string
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
      required:
        - name
        - grammy
    Album:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        name:
          type: string
          example: Innuendo
        year:
          type: number
          example: 1991
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
        - id
        - name
        - year
        - artistId
    AlbumDto:
      type: object
      properties:
        name:
          type: string
          example: Innuendo
        year:
          type: number
          example: 1991
        artistId:
          type: string
      required:
        - name
        - year
        - artistId
    Track:
      type: object
      properties:
        id:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        name:
          type: string
          example: The Show Must Go On
        duration:
          type: number
          example: 262
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        albumId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
        - id
        - name
        - duration
        - artistId
        - albumId
    TrackDto:
      type: object
      properties:
        name:
          type: string
          example: The Show Must Go On
        duration:
          type: number
          example: 262
          description: in seconds
        artistId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        albumId:
          type: string
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
        - name
        - duration
        - artistId
        - albumId